pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "source-atlas"
version = "0.1.0"
description = "Multi-language code analyzer with LSP and Neo4j integration"
authors = [
    {name = "Nguyen Van Quyen", email = "quyennv.4work@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["code-analysis", "ast", "lsp", "neo4j", "graph", "static-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "loguru>=0.7.3,<0.8.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "cryptography>=41.0.0,<43",
    "tree-sitter==0.25.1",
    "tree-sitter-language-pack==0.9.0",
    "jedi-language-server==0.41.1",
    "setuptools",
    "psutil>=7.0.0,<8.0.0",
    "requests==2.32.3",
    "neo4j==5.14.0",
    "pydantic-settings>=2.11.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/quyen-ngv/source-atlas"
Documentation = "https://github.com/quyen-ngv/source-atlas#readme"
Repository = "https://github.com/quyen-ngv/source-atlas"
Issues = "https://github.com/quyen-ngv/source-atlas/issues"
Changelog = "https://github.com/quyen-ngv/source-atlas/blob/main/CHANGELOG.md"

[project.scripts]
source-atlas = "source_atlas.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["source_atlas*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
source_atlas = ["py.typed"]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=source_atlas --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["source_atlas"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
